# frontend/src/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "LifeGrid API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/forecasts": {
   "get": {
    "tags": [
     "forecasting"
    ],
    "summary": "Get Series",
    "description": "28 days of actual usage plus the latest 7-day forecast for one series. Plasma is ABO-only (use rh=pos).",
    "operationId": "get_series_api_v1_forecasts_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Site Id"
      }
     },
     {
      "name": "component",
      "in": "query",
      "required": true,
      "schema": {
       "$ref": "#/components/schemas/Component"
      }
     },
     {
      "name": "abo",
      "in": "query",
      "required": true,
      "schema": {
       "$ref": "#/components/schemas/Abo"
      }
     },
     {
      "name": "rh",
      "in": "query",
      "required": false,
      "schema": {
       "$ref": "#/components/schemas/Rh",
       "default": "pos"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Series"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/forecasts/accuracy": {
   "get": {
    "tags": [
     "forecasting"
    ],
    "summary": "Get Accuracy",
    "operationId": "get_accuracy_api_v1_forecasts_accuracy_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Site Id"
      }
     },
     {
      "name": "component",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/Component"
        },
        {
         "type": "null"
        }
       ],
       "title": "Component"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Accuracy"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/donor/profile": {
   "get": {
    "tags": [
     "donors"
    ],
    "summary": "Get Profile",
    "operationId": "get_profile_api_v1_donor_profile_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DonorOut"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "donors"
    ],
    "summary": "Put Profile",
    "operationId": "put_profile_api_v1_donor_profile_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProfileIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DonorOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "donors"
    ],
    "summary": "Delete",
    "operationId": "delete_api_v1_donor_profile_delete",
    "responses": {
     "204": {
      "description": "Successful Response"
     }
    }
   }
  },
...
  "/api/v1/donors": {
   "get": {
    "tags": [
     "donors"
    ],
    "summary": "Search",
    "operationId": "search_api_v1_donors_get",
    "parameters": [
     {
      "name": "group",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(O|A|B|AB)(pos|neg)$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Group"
      }
     },
     {
      "name": "eligible",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "pattern": "^(yes|no)$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Eligible"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 40
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/CoordinatorDonorOut"
         },
         "title": "Response Search Api V1 Donors Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/donors/{donor_id}/deferrals": {
   "post": {
    "tags": [
     "donors"
    ],
    "summary": "Defer",
    "operationId": "defer_api_v1_donors__donor_id__deferrals_post",
    "parameters": [
     {
      "name": "donor_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Donor Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DeferralIn"
       }
      }
     }
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/donors/{donor_id}/verify-group": {
   "post": {
    "tags": [
     "donors"
    ],
    "summary": "Verify Group",
    "operationId": "verify_group_api_v1_donors__donor_id__verify_group_post",
    "parameters": [
     {
      "name": "donor_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Donor Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "Abo": {
    "type": "string",
    "enum": [
     "O",
     "A",
     "B",
     "AB"
    ],
    "title": "Abo"
   },
...
   "Accuracy": {
    "properties": {
     "days": {
      "type": "integer",
      "title": "Days"
     },
     "scored": {
      "type": "integer",
      "title": "Scored"
     },
     "mae": {
      "type": "number",
      "title": "Mae"
     },
     "coverage": {
      "type": "number",
      "title": "Coverage"
     }
    },
    "type": "object",
    "required": [
     "days",
     "scored",
     "mae",
     "coverage"
    ],
    "title": "Accuracy"
   },
...
   "Availability": {
    "type": "string",
    "enum": [
     "available",
     "unavailable",
     "travelling"
    ],
    "title": "Availability"
   },
...
   "Component": {
    "type": "string",
    "enum": [
     "red_cells",
     "platelets",
     "plasma"
    ],
    "title": "Component"
   },
   "CoordinatorDonorOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "ref": {
      "type": "string",
      "title": "Ref"
     },
     "abo": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Abo"
       },
       {
        "type": "null"
       }
      ]
     },
     "rh": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Rh"
       },
       {
        "type": "null"
       }
      ]
     },
     "group_verified": {
      "type": "boolean",
      "title": "Group Verified"
     },
     "area_label": {
      "type": "string",
      "title": "Area Label"
     },
     "last_donation_on": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Donation On"
     },
     "availability": {
      "$ref": "#/components/schemas/Availability"
     },
     "reliability": {
      "type": "number",
      "title": "Reliability"
     },
     "deferral": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Deferral"
       },
       {
        "type": "null"
       }
      ]
     },
     "eligible_from": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Eligible From"
     },
     "weight_ok": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Weight Ok"
     },
     "quiet_start": {
      "anyOf": [
       {
        "type": "string",
        "format": "time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quiet Start"
     },
     "quiet_end": {
      "anyOf": [
       {
        "type": "string",
        "format": "time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quiet End"
     },
     "emergency_override": {
      "type": "boolean",
      "title": "Emergency Override"
     },
     "max_invites_30d": {
      "type": "integer",
      "title": "Max Invites 30D"
     },
     "consent_version": {
      "type": "string",
      "title": "Consent Version"
     },
     "consent_at": {
      "type": "string",
      "format": "date-time",
      "title": "Consent At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "ref",
     "abo",
     "rh",
     "group_verified",
     "area_label",
     "last_donation_on",
     "availability",
     "reliability",
     "deferral",
     "eligible_from",
     "weight_ok",
     "quiet_start",
     "quiet_end",
     "emergency_override",
     "max_invites_30d",
     "consent_version",
     "consent_at"
    ],
    "title": "CoordinatorDonorOut",
    "description": "No contact details, no exact location (SEC-08, SEC-09)."
   },
...
   "Deferral": {
    "properties": {
     "category": {
      "type": "string",
      "title": "Category"
     },
     "eligible_again_on": {
      "type": "string",
      "format": "date",
      "title": "Eligible Again On"
     }
    },
    "type": "object",
    "required": [
     "category",
     "eligible_again_on"
    ],
    "title": "Deferral"
   },
   "DeferralIn": {
    "properties": {
     "category": {
      "type": "string",
      "pattern": "^[a-z_]{3,40}$",
      "title": "Category"
     },
     "eligible_again_on": {
      "type": "string",
      "format": "date",
      "title": "Eligible Again On"
     }
    },
    "type": "object",
    "required": [
     "category",
     "eligible_again_on"
    ],
    "title": "DeferralIn"
   },
...
   "DonorOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "ref": {
      "type": "string",
      "title": "Ref"
     },
     "abo": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Abo"
       },
       {
        "type": "null"
       }
      ]
     },
     "rh": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Rh"
       },
       {
        "type": "null"
       }
      ]
     },
     "group_verified": {
      "type": "boolean",
      "title": "Group Verified"
     },
     "area_label": {
      "type": "string",
      "title": "Area Label"
     },
     "area_lat": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Area Lat"
     },
     "area_lng": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Area Lng"
     },
     "birth_year": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Birth Year"
     },
     "weight_ok": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Weight Ok"
     },
     "last_donation_on": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Donation On"
     },
     "availability": {
      "$ref": "#/components/schemas/Availability"
     },
     "reliability": {
      "type": "number",
      "title": "Reliability"
     },
     "deferral": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Deferral"
       },
       {
        "type": "null"
       }
      ]
     },
     "eligible_from": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Eligible From"
     },
     "quiet_start": {
      "anyOf": [
       {
        "type": "string",
        "format": "time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quiet Start"
     },
     "quiet_end": {
      "anyOf": [
       {
        "type": "string",
        "format": "time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quiet End"
     },
     "emergency_override": {
      "type": "boolean",
      "title": "Emergency Override"
     },
     "max_invites_30d": {
      "type": "integer",
      "title": "Max Invites 30D"
     },
     "consent_version": {
      "type": "string",
      "title": "Consent Version"
     },
     "consent_at": {
      "type": "string",
      "format": "date-time",
      "title": "Consent At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "ref",
     "abo",
     "rh",
     "group_verified",
     "area_label",
     "area_lat",
     "area_lng",
     "birth_year",
     "weight_ok",
     "last_donation_on",
     "availability",
     "reliability",
     "deferral",
     "eligible_from",
     "quiet_start",
     "quiet_end",
     "emergency_override",
     "max_invites_30d",
     "consent_version",
     "consent_at"
    ],
    "title": "DonorOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Point": {
    "properties": {
     "day": {
      "type": "string",
      "format": "date",
      "title": "Day"
     },
     "actual": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Actual"
     },
     "point": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Point"
     },
     "low": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Low"
     },
     "high": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "High"
     }
    },
    "type": "object",
    "required": [
     "day"
    ],
    "title": "Point"
   },
   "ProfileIn": {
    "properties": {
     "abo": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Abo"
       },
       {
        "type": "null"
       }
      ]
     },
     "rh": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Rh"
       },
       {
        "type": "null"
       }
      ]
     },
     "area_lat": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 90.0,
        "minimum": -90.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Area Lat"
     },
     "area_lng": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 180.0,
        "minimum": -180.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Area Lng"
     },
     "birth_year": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 2100.0,
        "minimum": 1900.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Birth Year"
     },
     "weight_ok": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Weight Ok"
     },
     "availability": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Availability"
       },
       {
        "type": "null"
       }
      ]
     },
     "quiet_start": {
      "anyOf": [
       {
        "type": "string",
        "format": "time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quiet Start"
     },
     "quiet_end": {
      "anyOf": [
       {
        "type": "string",
        "format": "time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quiet End"
     },
     "emergency_override": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Emergency Override"
     },
     "max_invites_30d": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 10.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Invites 30D"
     },
     "consent": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Consent"
     },
     "consent_version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Consent Version"
     }
    },
    "type": "object",
    "title": "ProfileIn"
   },
...
   "Rh": {
    "type": "string",
    "enum": [
     "pos",
     "neg"
    ],
    "title": "Rh"
   },
...
   "Series": {
    "properties": {
     "model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     },
     "model_version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model Version"
     },
     "mae": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mae"
     },
     "baseline_mae": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Baseline Mae"
     },
     "coverage": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Coverage"
     },
     "points": {
      "items": {
       "$ref": "#/components/schemas/Point"
      },
      "type": "array",
      "title": "Points"
     }
    },
    "type": "object",
    "required": [
     "model",
     "model_version",
     "mae",
     "baseline_mae",
     "coverage",
     "points"
    ],
    "title": "Series"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}